# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  loggedIn: boolean;
  /** @nullable */
  userID: string | null;
  /** @nullable */
  userName: string | null;
  uptime: number;
  reconnectAttempts: number;
  /** @nullable */
  lastUpdated: string | null;
  status: string;
}

export type CookiesInputCookiesItem = { [key: string]: unknown };

export interface CookiesInput {
  cookies: CookiesInputCookiesItem[];
}

export interface UpdateResult {
  success: boolean;
  message: string;
}

export interface BotCommand {
  name: string;
  enabled: boolean;
  description: string;
}

export interface ErrorResponse {
  error: string;
}